refactor(web): extract MediaContent from media-sidebar into media-content.tsx
Moves MediaContent, LiveStreamSection, and MediaItemCard into a dedicated
media-content.tsx file. media-sidebar.tsx drops from 673 → 344 lines and
now only contains the sidebar shell (tab header, tab switching, pinned/drawer
wrapper). No behavior or UI changes.
